requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic-settings
python-dotenv
docker
jinja2
python-jose[cryptography]
